requirements.txt
amqp==5.0.9
asgiref==3.4.1
AWSIoTPythonSDK==1.4.9
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.1
Django==3.2.11
dlib==19.22.1
dnspython==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.2
importlib-metadata==4.10.1
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
numpy==1.21.5
opencv-python==4.5.5.62
paho-mqtt==1.6.1
Pillow==9.0.0
prompt-toolkit==3.0.24
pygame==2.1.2
pyngrok==5.1.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.7.0
